requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
chardet
openpyxl
scipy
wordcloud
msoffcrypto-tool
xlrd
